pyproject.toml
[tool.poetry]
name = "vibetm-be"
version = "0.1.0"
description = ""
authors = ["irakli <samsaxuri123@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.1.3"
flask-sqlalchemy = "^3.1.1"
flask-bcrypt = "^1.0.1"
pyjwt = "^2.11.0"
python-dotenv = "^1.2.1"
psycopg2-binary = "^2.9.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
